improved manipulation of images using the position plugin

+++ Plugins/org.mitk.gui.qt.m2.Position/src/internal/m2Position.h
@@ -51,7 +51,7 @@ protected:
                                   const QList<mitk::DataNode::Pointer> &nodes) override;
 
   /// \brief Called when the user clicks the GUI button
-  void Rotate(int angleDeg);
+  void Rotate(float angleDeg);
   void Mirror(int angleDeg);
   void Move(std::array<int,2> && vec);
